+++ src/shared/filepic.c
@@ -8,7 +8,7 @@
 
 int openFile(const char *name, int mode);
 void fileClose(int handle);
-void decodePic(int handle, int segment);
+void decodePic(int handle, uint16 segment);
 
 void mystrcpy(char *dest, const char *source) {
     do {
@@ -46,7 +46,7 @@ void openShowPic(char *filename, int page, int garbage) /* Original chain: OpenF
     TRACE(("openShowPic: file closed, returning"));
 }
 
-void loadPic(const char *filename, int segment) { /* Original chain: OpenFile + DecodePic(InSeg, OutSeg) + CloseFile. Load PIC into segment. */
+void loadPic(const char *filename, uint16 segment) { /* Original chain: OpenFile + DecodePic(InSeg, OutSeg) + CloseFile. Load PIC into segment. */
     int handle;
     handle = openFileWrapper(filename, 0);
     TRACE(("loadPic(): opened %s, loading into segment 0x%x", filename, segment));
